requirements.txt
torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
mrcfile>=1.4.0
pyyaml>=6.0
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.10.0
